move jobs executions management to ScheduledJobManager
